Stop using gradle plugin..

+++ idsBuild.sh
@@ -18,14 +18,39 @@ echo Ca Cert `echo $CA_CERT | md5sum`
 echo Cert `echo $CERT | md5sum`
 cd ..
 
+echo Obtaining docker.
+wget http://security.ubuntu.com/ubuntu/pool/main/a/apparmor/libapparmor1_2.8.95-2430-0ubuntu5.3_amd64.deb -O libapparmor.deb
+sudo dpkg -i libapparmor.deb 
+rm libapparmor.deb
+wget https://get.docker.com/builds/Linux/x86_64/docker-1.9.1 --quiet -O docker
+chmod +x docker
+
 echo Building projects using gradle...
 ./gradlew build
-echo Building and Starting Map Docker Image...
-cd map-wlpcfg
-../gradlew buildDockerImage
-../gradlew stopCurrentContainer
-../gradlew removeCurrentContainer
-../gradlew startNewEtcdContainer
-
-cd ..
-rm -rf dockercfg
+if [ $? != 0 ]
+then
+  echo Gradle Build failed.
+  exit -1
+else
+  cd map-wlpcfg
+  ../docker build -t gameon-map
+  if [ $? != 0 ]
+  then 
+    echo Docker Build failed.
+    exit -2
+  else
+    echo Attempting to remove old containers.
+    ../docker stop -t 0 gameon-map || true
+    ../docker rm gameon-map || true
+    echo Starting new container.
+    ../docker run -d -p 9087:9080 -p 9447:9443 --link etcd -e LICENSE=accept -e ETCDCTL_ENDPOINT=http://etcd:4001 --name=gameon-map gameon-map
+    if [ $? != 0 ]
+    then
+      echo Docker Run failed.
+      exit -3
+    else
+      cd ..
+      rm -rf dockercfg
+    fi
+  fi
+fi
